requirements.txt
streamlit
psycopg2-binary
pandas
plotly
seaborn
matplotlib
streamlit-autorefresh